Restructured layout for first project & added proper description
This is going to be the new structure for the rest of the projects. I did not have room for the third paragraph, I need to change how the project detail sections open to allow easy vertical scrolling. Currently vertical scrolling triggers horizontal movement between all main sections, perhaps this could be changed.